Falta BD

+++ horario/models.py
@@ -5,7 +5,7 @@ class Horario(models.Model):
     id = models.AutoField(primary_key=True)
     sigla = models.CharField(max_length=20, blank=False)
     asignatura = models.TextField(blank=False)
-    paralelo = models.IntegerField(blank=False)
+    paralelo = models.PositiveSmallIntegerField(blank=False)
     horario = models.TextField()
 
     def __str__(self):
@@ -14,7 +14,7 @@ class Horario(models.Model):
 class Opcion(models.Model):
     id = models.AutoField(primary_key=True)
     carrera = models.CharField(max_length=200,blank=False)
-    opcion = models.IntegerField(blank=False)
+    opcion = models.PositiveSmallIntegerField(blank=False)
     horario = models.ManyToManyField(Horario)
 
     def __str__(self):
